Move Configuration into Abstractions Project
