+++ NuKeeper.Integration.Tests/Engine/RepositoryFilterTests.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Threading.Tasks;
 using NSubstitute;
+using NuKeeper.Abstractions.Configuration;
 using NuKeeper.Configuration;
 using NuKeeper.Engine;
 using NuKeeper.GitHub;
@@ -42,7 +43,7 @@ namespace NuKeeper.Integration.Tests.Engine
             var logger = Substitute.For<INuKeeperLogger>();
 
             var gitHubClient = new OctokitClient(logger);
-            gitHubClient.Initialise(new GithubAuthSettings(new Uri("https://api.github.com"), testKeyWithOnlyPublicAccess));
+            gitHubClient.Initialise(new AuthSettings(new Uri("https://api.github.com"), testKeyWithOnlyPublicAccess));
 
             return new RepositoryFilter(gitHubClient, logger);
         }
