+++ NuKeeper.Tests/Commands/InspectCommandTests.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Threading.Tasks;
 using NSubstitute;
+using NuKeeper.Abstractions.Configuration;
 using NuKeeper.Commands;
 using NuKeeper.Configuration;
 using NuKeeper.Inspection.Logging;
