+++ NuKeeper.Tests/Engine/ForkFinderTests.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Threading.Tasks;
 using NSubstitute;
+using NuKeeper.Abstractions.Configuration;
 using NuKeeper.Configuration;
 using NuKeeper.Engine;
 using NuKeeper.GitHub;
