+++ NuKeeper.sln
@@ -20,7 +20,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NuKeeper.Inspection.Tests",
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NuKeeper.Update", "NuKeeper.Update\NuKeeper.Update.csproj", "{E210D6EF-4EC3-44DD-8E18-269A3647688B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NuKeeper.Update.Tests", "NuKeeper.Update.Tests\NuKeeper.Update.Tests.csproj", "{1336C9CF-8B90-428B-904B-2C223010FD16}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NuKeeper.Update.Tests", "NuKeeper.Update.Tests\NuKeeper.Update.Tests.csproj", "{1336C9CF-8B90-428B-904B-2C223010FD16}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NuKeeper.Abstractions", "NuKeeper.Abstractions\NuKeeper.Abstractions.csproj", "{CE51CCF6-35BA-4FF5-A44C-D348D989E11E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -56,6 +58,10 @@ Global
 		{1336C9CF-8B90-428B-904B-2C223010FD16}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1336C9CF-8B90-428B-904B-2C223010FD16}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1336C9CF-8B90-428B-904B-2C223010FD16}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CE51CCF6-35BA-4FF5-A44C-D348D989E11E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CE51CCF6-35BA-4FF5-A44C-D348D989E11E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CE51CCF6-35BA-4FF5-A44C-D348D989E11E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CE51CCF6-35BA-4FF5-A44C-D348D989E11E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
