+++ NuKeeper/Commands/LocalNuKeeperCommand.cs
@@ -2,6 +2,7 @@ using McMaster.Extensions.CommandLineUtils;
 using NuKeeper.Configuration;
 using NuKeeper.Inspection.Logging;
 using System.IO;
+using NuKeeper.Abstractions.Configuration;
 
 namespace NuKeeper.Commands
 {
