+++ NuKeeper/Commands/MultipleRepositoryCommand.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Text.RegularExpressions;
 using McMaster.Extensions.CommandLineUtils;
+using NuKeeper.Abstractions.Configuration;
 using NuKeeper.Configuration;
 using NuKeeper.Engine;
 using NuKeeper.Inspection.Logging;
