+++ NuKeeper/Configuration/SettingsContainer.cs
@@ -1,3 +1,4 @@
+using NuKeeper.Abstractions.Configuration;
 using NuKeeper.Update.Selection;
 
 namespace NuKeeper.Configuration
@@ -6,7 +7,7 @@ namespace NuKeeper.Configuration
     {
         public SourceControlServerSettings SourceControlServerSettings { get; set; }
 
-        public GithubAuthSettings GithubAuthSettings { get; set; }
+        public AuthSettings GithubAuthSettings { get; set; }
 
         public UserSettings UserSettings { get; set; }
