+++ NuKeeper/Engine/IForkFinder.cs
@@ -1,4 +1,5 @@
 using System.Threading.Tasks;
+using NuKeeper.Abstractions.Configuration;
 using NuKeeper.Configuration;
 
 namespace NuKeeper.Engine
