+++ NuKeeper/GitHub/IGitHub.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using NuKeeper.Configuration;
+using NuKeeper.Abstractions.Configuration;
 using NuKeeper.Engine;
 using Octokit;
 
@@ -8,7 +8,7 @@ namespace NuKeeper.GitHub
 {
     public interface IGitHub
     {
-        void Initialise(GithubAuthSettings settings);
+        void Initialise(AuthSettings settings);
 
         Task<Account> GetCurrentUser();
